Update goja, peg js deps to specific version
Later versions start to degrade test performance quite a bit, and
latest versions flat out do not work. This seems to be due to some
weirdness in the require() logic where a circular import does not
work correctly. There is an existing circular import in regenerator-
transform (visit -> emit -> visit) so it works in that example but
not others.